move ContextEffect.boundary to separate Boundary object + fix test class name for ContextEffect
